+++ src/components/event/forms/index.ts
@@ -3,3 +3,4 @@ export { DoctorVisitForm, createEmptyDoctorVisit } from './DoctorVisitForm';
 export { InterventionForm, createEmptyIntervention } from './InterventionForm';
 export { MedicationForm, createEmptyMedication } from './MedicationForm';
 export { LabResultForm, createEmptyLabResult } from './LabResultForm';
+export { ViceForm, createEmptyVice } from './ViceForm';